{"schema":"decodex/commit/1","summary":"Extract overlay image sampler","authority":"manual"}
